requirements.txt
rpi.gpio
python-dotenv
requests